Convert ARM optimizations to GNU syntax
The ARM optimizations included in OPUS 1.1 is written in ARM's RVCT syntax, and we need to convert it to GNU syntax. There's a script for the actual conversion, arm2gnu.pl, added by this CL.

The script will be used in opus.gyp to generate files of the format we need.

BUG=
R=sergeyu@chromium.org

Review URL: https://codereview.chromium.org/159803003

git-svn-id: svn://svn.chromium.org/chrome/trunk/deps/third_party/opus@256783 0039d316-1c4b-4281-b951-d872f2087c98